+++ test/schema_parser_test.py
@@ -7,7 +7,7 @@ import copy
 
 import pytest
 from marshmallow.exceptions import MarshmallowError
-from pytest_lazyfixture import lazy_fixture
+from pytest_lazy_fixtures import lf as lazy_fixture
 
 import brewtils.models
 from brewtils.models import System
@@ -59,12 +59,6 @@ class TestParse(object):
     def test_error(self, data, kwargs, error):
         with pytest.raises(error):
             SchemaParser.parse_system(data, **kwargs)
-
-    def test_non_strict_failure(self, system_dict):
-        system_dict["name"] = 1234
-        value = SchemaParser.parse_system(system_dict, from_string=False, strict=False)
-        assert value.get("name") is None
-        assert value["version"] == system_dict["version"]
 
     def test_no_modify(self, system_dict):
         system_copy = copy.deepcopy(system_dict)